extract lineup
