+++ package.json
@@ -27,6 +27,7 @@
     "react-scripts": "0.7.0"
   },
   "dependencies": {
+    "jquery": "^3.1.1",
     "react": "^15.3.2",
     "react-dom": "^15.3.2",
     "showdown": "^1.5.0"
